A bit more doctor refactor
